Implement q_delete_dup
Use two pointers to iterate through the list and detect the duplicate
string. When detecting the duplicate string, use one pointer points to
the head of the sublist and use another pointer points to the tail of it.
Finally, Move the sublist to another temporary list head and use `q_free()`
to delete the nodes.