Implement q_ascend and q_descend
Because the implementations of q_ascend() and q_descend() are nearly
identical, except for the comparision between node values, I create
another function q_descend_or_ascend() to handle them simultaneously.

Iterate through the list from the tail of the list to head and use two
pointers point to two nodes for comparing values. If the list has to
be ascending and the previous node value is greater than the next
node, remove the previous one. If the list has to be descending and
the previous node value is less than the next node, also remove the
previous one.