Analyze the performance of Merge sort & list_sort
Use Python to execute `perf` for analyzing CPU cycles and cache
misses of Merge sort and list_sort. It tests multiple cases of
number of nodes for sorting, from 1000 to 500000, with intervals
of 1000. Finally, plot the Line graph for CPU cycles and cache
misses events to compare the performance of Merge sort and
list_sort.